# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aikya Sentinel – Unified AI Financial Crime Intelligence & Insider Threat Early Warning System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Employees
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Accounts
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: account_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found

  # Alerts
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
    Employee:
      type: object
      required: [id, employeeId, department, designation, trustScore, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: string
        department:
          type: string
        designation:
          type: string
        trustScore:
          type: number
        riskLevel:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [employeeId, department, designation]
      properties:
        employeeId:
          type: string
        department:
          type: string
        designation:
          type: string
        trustScore:
          type: number
    EmployeeUpdate:
      type: object
      properties:
        department:
          type: string
        designation:
          type: string
        trustScore:
          type: number
        riskLevel:
          type: string
...
    Transaction:
      type: object
      required: [id, transactionId, sourceAccount, destinationAccount, amount, timestamp, status]
      properties:
        id:
          type: integer
        transactionId:
          type: string
        sourceAccount:
          type: string
        destinationAccount:
          type: string
        amount:
          type: number
        currency:
          type: ["string", "null"]
        transactionType:
          type: ["string", "null"]
        status:
          type: string
        flagged:
          type: boolean
        timestamp:
          type: string
    TransactionInput:
      type: object
      required: [transactionId, sourceAccount, destinationAccount, amount]
      properties:
        transactionId:
          type: string
        sourceAccount:
          type: string
        destinationAccount:
          type: string
        amount:
          type: number
        currency:
          type: string
        transactionType:
          type: string
...